backend/pyproject.toml
[project]
name = "liriac-backend"
version = "0.1.0"
description = "Local Django backend for Liriac"
authors = [{ name = "Liriac" }]
requires-python = ">=3.11"
dependencies = [
    "django>=5.1,<6",
    "djangorestframework>=3.15,<4",
    "drf-spectacular>=0.27,<0.28",
    "django-cors-headers>=4.4,<5",
]

[project.optional-dependencies]
dev = [
    "pytest-django>=4.8,<5",
    "ruff>=0.6,<0.7"
]

[tool.uv]
package = true

[tool.setuptools.packages.find]
include = ["config", "studio"]
